Moved get_default_finder into finders module
